Logger

+++ TM_CommonPy/_Logger.py
@@ -1,6 +1,5 @@
 import os
 import logging
-import TM_CommonPy as TM
 # Settings
 bWriteLogFile = True
 sLogFile = os.path.join(__file__, '..', 'TMLog.log')
@@ -11,26 +10,20 @@ vFileHandlerThreshold = logging.DEBUG
 TMLog = logging.getLogger(__name__)
 TMLog.setLevel(vMasterThreshold)
 vFormatter = logging.Formatter('%(levelname)-7s %(message)s')
-# ---ConsoleHandler
-vConsoleHandler = logging.StreamHandler()
-vConsoleHandler.setLevel(vConsoleHandlerThreshold)
-vConsoleHandler.setFormatter(vFormatter)
-TMLog.addHandler(vConsoleHandler)
-# ---FileHandler
-try:
-    os.remove(sLogFile)
-except (PermissionError, FileNotFoundError):
-    pass
-if bWriteLogFile:
-    bLogFileIsOpen = False
+# Handlers
+if len(logging.getLogger().handlers):  # Default handlers are empty when it's the weird nosetests extra run.
+    # ---FileHandler
     try:
-        os.rename(sLogFile, sLogFile)
-    except PermissionError:
-        bLogFileIsOpen = True
+        os.remove(sLogFile)
     except FileNotFoundError:
         pass
-    if not bLogFileIsOpen:
+    if bWriteLogFile:
         vFileHandler = logging.FileHandler(sLogFile)
         vFileHandler.setFormatter(vFormatter)
         vFileHandler.setLevel(vFileHandlerThreshold)
         TMLog.addHandler(vFileHandler)
+    # ---ConsoleHandler
+    vConsoleHandler = logging.StreamHandler()
+    vConsoleHandler.setLevel(vConsoleHandlerThreshold)
+    vConsoleHandler.setFormatter(vFormatter)
+    TMLog.addHandler(vConsoleHandler)
